+++ src/app/app-routing.module.ts
@@ -2,21 +2,26 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { AuthGuard } from './shared/guards/auth-guard/auth.guard';
 
+/**
+ * Adjust page width is done in component's template.
+ */
+
 const routes: Routes = [
   {
     path: '',
     loadChildren: () =>
-      import('./search-module/search.module').then((module) => module.SearchModule),
+      import('./modules/search-module/search.module').then((module) => module.SearchModule),
   },
   {
     path: 'admin',
-    loadChildren: () => import('./admin-module/admin.module').then((module) => module.AdminModule),
+    loadChildren: () =>
+      import('./modules/admin-module/admin.module').then((module) => module.AdminModule),
     canActivate: [AuthGuard],
   },
   {
     path: '**',
     loadComponent: () =>
-      import('./not-found-page/not-found-page.component').then(
+      import('./modules/not-found-page/not-found-page.component').then(
         (module) => module.NotFoundPageComponent
       ),
   },