+++ src/app/app.module.ts
@@ -6,7 +6,8 @@ import { RECAPTCHA_V3_SITE_KEY, RecaptchaV3Module } from 'ng-recaptcha';
 import { environment } from '../environments/environment';
 import { AppComponent } from './app.component';
 import { AppRoutingModule } from './app-routing.module';
-import { LocalizationModule } from './localization-module/localization.module';
+import { LocalizationModule } from './modules/localization-module/localization.module';
+import { LoaderComponent } from "./reusable-component/loader/loader.component";
 import { HttpRequestInterceptor } from './shared/services/http-request-interceptor/http-request.interceptor';
 
 @NgModule({
@@ -18,6 +19,7 @@ import { HttpRequestInterceptor } from './shared/services/http-request-intercept
     AppRoutingModule,
     LocalizationModule.forRoot(),
     RecaptchaV3Module,
+    LoaderComponent,
   ],
   providers: [
     {