+++ src/app/app.settings.ts
@@ -1,4 +1,3 @@
-
 export const APPLICATION_NAME = 'Recipier';
 
 const SEARCH_PAGE_TITLE = {
@@ -8,11 +7,11 @@ const SEARCH_PAGE_TITLE = {
 const PAGE_NOT_FOUND = {
   cs: `Stránka nenalezena | ${APPLICATION_NAME}`,
   en: `Page Not Found | ${APPLICATION_NAME}`,
-}
+};
 
 export const APP_PAGE_TITLE = {
   SEARCH: SEARCH_PAGE_TITLE,
-  PAGE_NOT_FOUND: PAGE_NOT_FOUND
+  PAGE_NOT_FOUND: PAGE_NOT_FOUND,
 };
 
 // TODO add meta tags