+++ src/app/modules/localization-module/services/general-locale.token.ts
@@ -1,0 +1,4 @@
+import { InjectionToken } from '@angular/core';
+import { LocaleFileKey } from '../models/locale-file-key.type';
+
+export const GeneralLocale = new InjectionToken<LocaleFileKey>('General Locale Texts');