+++ src/app/shared/models/response-list.type.ts
@@ -1,4 +1,4 @@
 export type ResponseList<T> = {
   list: Array<T>;
   length: number;
-}
+};